+++ sw.js
@@ -1,4 +1,4 @@
-const CACHE_NAME = "ihnatiev-site-v0.6.22b-r1";
+const CACHE_NAME = "ihnatiev-site-v0.6.23a-r1";
 const APP_SHELL = [
   "./",
   "./index.html",
@@ -24,6 +24,7 @@ const APP_SHELL = [
   "./css/responsive.css",
   "./css/styles.css",
   "./config.js",
+  "./files/content/home.json",
   "./js/site-utils.js",
   "./js/gallery-lightbox.js",
   "./js/gallery-renderer.js",